LayoutTests/imported/w3c/web-platform-tests/fetch/http-cache/partial-expected.txt


FAIL HTTP cache stores partial content and reuses it. assert_equals: Response body expected "3e5b8a86-d5f7-419c-9232-2ae3ac5941ab" but got "01234"
FAIL HTTP cache stores complete response and serves smaller ranges from it. assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores partial response and serves smaller ranges from it. assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores partial content and completes it. assert_equals: expected (string) "bytes=5-" but got (undefined) undefined
